Move `alias this` directly into GodotScript
DMD v2.084 fixes many `alias this` bugs, including it not being inherited by subclasses. It no longer needs to be added to script classes manually, so this is no longer mentioned in documentation.

The test project needs to be updated once LDC also has this fix. Any additional `alias this` bug workarounds can be removed at that point as well.